Separate application from library
