+++ src/signal-server.hh
@@ -30,5 +30,8 @@ void alloc_elev(void);
 void alloc_path(void);
 void alloc_dem(void);
 void do_allocs(void);
+int handle_args(int argc, char *argv[]);
+int scan_stdin();
+const char *version();
 
 #endif /* _MAIN_HH_ */
